# api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/plaid/me/auth": {
      "get": {
        "description": "Calls /auth/get for each Item. Requires the auth product to be enabled on the Item. In sandbox the default test institution returns deterministic numbers.",
        "operationId": "PlaidMeController_getMyAuth",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Array of per-item auth payloads (accounts + numbers)"
          },
          "403": {
            "description": "No Plaid account linked"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get routing and account numbers for the user's linked accounts",
        "tags": ["plaid"]
      }
    },
...
    "/auth/register": {
      "post": {
        "operationId": "AuthController_register",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User created successfully"
          },
          "409": {
            "description": "Username already taken"
          }
        },
        "summary": "Register a new user",
        "tags": ["auth"]
      }
    },
    "/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Returns accessToken and user profile"
          },
          "401": {
            "description": "Invalid credentials"
          }
        },
        "summary": "Log in; sets httpOnly cookies and also returns accessToken",
        "tags": ["auth"]
      }
    },
    "/auth/refresh": {
      "post": {
        "operationId": "AuthController_refresh",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Returns new accessToken"
          },
          "401": {
            "description": "Invalid or expired refresh token"
          }
        },
        "summary": "Rotate refresh token (reads cookie or body) and issue a new access token",
        "tags": ["auth"]
      }
    },
...
    "/auth/me": {
      "get": {
        "operationId": "AuthController_me",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Current user profile from JWT"
          }
        },
        "security": [
          {
            "cookie": []
          },
          {
            "bearer": []
          }
        ],
        "summary": "Get the currently authenticated user",
        "tags": ["auth"]
      }
    },
...
  },
  "info": {
    "title": "Flamingo API",
    "description": "Flamingo REST API",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      },
      "admin-jwt": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "RegisterDto": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "example": "alice",
            "minLength": 3,
            "maxLength": 150
          },
          "password": {
            "type": "string",
            "example": "S3cure!Pass",
            "minLength": 8
          },
          "email": {
            "type": "string",
            "example": "alice@example.com"
          }
        },
        "required": ["username", "password"]
      },
      "LoginDto": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "example": "alice"
          },
          "password": {
            "type": "string",
            "example": "S3cure!Pass"
          }
        },
        "required": ["username", "password"]
      },
...
    }
  }
}